backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.0
pinecone-client==3.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
aiofiles==23.2.1
httpx==0.25.0
pydantic-core==2.13.0
